A change to AKOperations use of Numerics caused many things to compile in Swift that should have generated warnings
I suspect quite a few clients will notice stuff like this switching from AK4 to 5

+++ Sources/AudioKit/User Interface/AKKeyboardView.swift
@@ -522,7 +522,8 @@ public class AKKeyboardView: NSView, AKMIDIListener {
         self.octaveCount = octaveCount
         self.firstOctave = firstOctave
         super.init(frame: CGRect(width: width, height: height))
-        size = CGSize(width: width / octaveCount - width / (octaveCount * octaveCount * 7), height: Double(height))
+        size = CGSize(width: Double(width) / Double(octaveCount) - Double(width) / Double(octaveCount * octaveCount * 7),
+                      height: Double(height))
         needsDisplay = true
     }
